// src/types/graphql.d.ts
export type Maybe<T> = T | null;
export type FieldWrapper<T> = T;
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type IUser = {
  id: FieldWrapper<Scalars['ID']>;
  nickname: FieldWrapper<Scalars['String']>;
  avatar: FieldWrapper<Scalars['String']>;
};

export type IChannel = {
  id: FieldWrapper<Scalars['ID']>;
  name: FieldWrapper<Scalars['String']>;
  isFavorite: FieldWrapper<Scalars['Boolean']>;
  messages?: Maybe<Array<FieldWrapper<IMessage>>>;
};

export type ICreateChannelInput = {
  name: Scalars['String'];
};

export type IFavoriteChannelInput = {
  id: Scalars['ID'];
  isFavorite: Scalars['Boolean'];
};

export type IMessage = {
  id: FieldWrapper<Scalars['ID']>;
  type: FieldWrapper<Scalars['String']>;
  createdAt: FieldWrapper<Scalars['String']>;
};

export type ITextMessage = IMessage & {
  id: FieldWrapper<Scalars['ID']>;
  type: FieldWrapper<Scalars['String']>;
  createdAt: FieldWrapper<Scalars['String']>;
  user: FieldWrapper<IUser>;
  message: FieldWrapper<Scalars['String']>;
};

export type ISystemMessage = IMessage & {
  id: FieldWrapper<Scalars['ID']>;
  type: FieldWrapper<Scalars['String']>;
  createdAt: FieldWrapper<Scalars['String']>;
  title: FieldWrapper<Scalars['String']>;
  message: FieldWrapper<Scalars['String']>;
};

export type ISendMessageInput = {
  channelId: Scalars['ID'];
  message: Scalars['String'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;
// ...
export type IResolversTypes = {
  Query: ResolverTypeWrapper<{}>;
  Int: ResolverTypeWrapper<Scalars['Int']>;
  String: ResolverTypeWrapper<Scalars['String']>;
  Subscription: ResolverTypeWrapper<{}>;
  Mutation: ResolverTypeWrapper<{}>;
  User: ResolverTypeWrapper<IUser>;
  ID: ResolverTypeWrapper<Scalars['ID']>;
  Channel: ResolverTypeWrapper<IChannel>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  CreateChannelInput: ICreateChannelInput;
  FavoriteChannelInput: IFavoriteChannelInput;
  Message: IResolversTypes['TextMessage'] | IResolversTypes['SystemMessage'];
  TextMessage: ResolverTypeWrapper<ITextMessage>;
  SystemMessage: ResolverTypeWrapper<ISystemMessage>;
  SendMessageInput: ISendMessageInput;
};